Fix update awk script
The condition surrounding whether an entry should be active or not was
flawed. It did not match when the start time as later than the end time.
Now it does.

Also change the awk script' shebang, as the env version does not permit
to append arguments without a script wrapper. The lightest solution was
to use the "de facto" path of awk, /usr/bin/awk. Less portable than a
shell wrapper but simple enough to do the job.